blog_project/requirements.txt
certifi==2020.6.20
Django==2.1
djangorestframework==3.11.1
mysqlclient==2.0.1
pytz==2020.1
wincertstore==0.2
